Addressed all comments from @prodonjs: - Moved all workgroup related routes to /api/workgroup - Added a 404 to API - Reverted req.user.auth in favor of all headers (explained in PR)
